initial stream inference
