+++ src/configs.ini
@@ -1,0 +1,19 @@
+[RECORDING]
+record_args = dict(energy_threshold=1000, record_timeout=2.5, phrase_timeout=3)
+prediction_weights = "./weights/model_attention_asr.pt"
+path_to_banned_words = "./data/banned_words.txt"
+sample_rate = 16000
+
+[MODEL]
+# number of features (MFCC + text label)
+input_dim = 222
+hidden__dim = 1024
+num_layers = 2
+# number of classes ("silence", "neutral", "negative")
+output_dim = 3
+
+[DATASET]
+# number of MFCC bins. Defaults to 17.
+num_bins = 17
+# number of elements in sequence. Defaults to 7.
+num_elems = 7
